Move avatar related code from AuthService to AvatarService.
